packages/agents/crosscut/pyproject.toml
[project]
name = "crosscut"
version = "0.1.0"
description = "天気データ分析AIエージェント - 横断分析エージェント（Step 3）"
requires-python = ">=3.12"
dependencies = [
    "strands-agents>=0.1",
    "strands-agents-tools>=0.1",
    "bedrock-agentcore>=1.4",
    "shared",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/crosscut"]
